attempt login registration

+++ src/main/java/com/prosilion/scdecisionmatrix/repository/ContractRepository.java
@@ -10,5 +10,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface ContractRepository extends JpaRepository<Contract, Integer> {
 	@Query("SELECT c FROM Contract c WHERE c.contractUser.id = :userId")
-	List<Contract> getContracts(@NonNull Long userId);
+	List<Contract> getContracts(@NonNull Integer userId);
 }
